+++ beast-of-argh/CHANGELOG.md
@@ -10,6 +10,7 @@
 * Adjust json files for grafana charts to set null id
 * Adjust blackbox exporter config and sample targets file
 * Split blackbox exporter into multiple jobs for http_2xx, icmp and tcp_connect
+* Switch to grafana10 in the grafana package
 
 ---